# docs/swagger.yaml
basePath: /
definitions:
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  handler.signInInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  tc.User:
    properties:
      email:
        type: string
      module_name:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
    required:
    - email
    - name
    - password
    type: object
...
host: tc.kaztranscom.kz:9000
info:
  contact:
    email: 7773785631@mail.ru,K.Yatsenko@kaztranscom.kz
    name: Серышев А.В., Яценко К.Ю.
  description: API Сервер для Программы Total Control
  title: Total Control
  version: "1.0"
paths:
...
  /api/ad_log:
    get:
      consumes:
      - application/json
      description: Список данных AD log по переданным параметрам
      operationId: get-ad-log-by-query
      parameters:
      - description: в формате yyyy-mm
        in: query
        name: date
        type: string
      - description: начало поиска в формате yyyy-mm-dd,(включительно)
        in: query
        name: from_date
        type: string
      - description: конец поиска в формате yyyy-mm-dd,(включительно)
        in: query
        name: to_date
        type: string
      - description: не строгий поиск, логин или фамилия или департамент
        in: query
        name: ad_login
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Список данных Ad log
      tags:
      - Active Directory
  /api/ad_log/check_log_file_csv:
    get:
      consumes:
      - application/json
      description: |-
        Проверка на существование лог файла в папке, для загрузки в базу данных,
        Путь к папке в файле настроек проекта
      operationId: check-log-file-csv
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Проверка и загрузка данных
      tags:
      - Active Directory
  /api/ad_log/download:
    get:
      consumes:
      - application/json
      description: Скачивание файла Excell с данными лога
      operationId: download-excell-file-ad-log
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Скачивание файла
      tags:
      - Active Directory
...
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Авторизация, получение токена для закрытых функций Api
      operationId: login
      parameters:
      - description: credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Авторизация
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Добавление в базу нового пользователя
      operationId: create-account
      parameters:
      - description: Данные для создания
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/tc.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Создание пользователя
      tags:
      - auth
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"